zdc
zecds

+++ Code/graphic.cc
@@ -43,7 +43,7 @@ void graphic_draw_window()
     
 }
 
-void graphic_draw_carre_plein(double x,double y,double size, double r, double g, double b){
+void graphic_draw_carre_plein(unsigned int x,unsigned int y,double size, double r, double g, double b){
     if (size ==1){
         (*ptcr)->set_line_width(1);
         (*ptcr)->set_source_rgb(r, g, b);
@@ -69,7 +69,7 @@ void graphic_draw_carre_plein(double x,double y,double size, double r, double g,
         (*ptcr)->stroke();*/
     }
 }
-void graphic_draw_carre_vide(double x, double y, double size, double r, double g, double b){
+void graphic_draw_carre_vide(unsigned int x,unsigned int y, double size, double r, double g, double b){
     (*ptcr)->set_line_width(0.5);
     (*ptcr)->set_source_rgb(r, g, b);
     (*ptcr)->move_to(x, y-0.5);
@@ -82,14 +82,14 @@ void graphic_draw_carre_vide(double x, double y, double size, double r, double g
     (*ptcr)->line_to(x, y-0.5);
     (*ptcr)->stroke();
     }
-void graphic_draw_losange(double x,double y,double r, double g, double b){
+void graphic_draw_losange(unsigned int x,unsigned int y,double r, double g, double b){
     (*ptcr)->set_line_width(1./sqrt(2));
     (*ptcr)->set_source_rgb(r, g, b);
     (*ptcr)->move_to(x-0.25, y-0.25);
     (*ptcr)->line_to(x+0.25, y+0.25);
     (*ptcr)->stroke();
 }
-void graphic_draw_plus(double x,double y, double size, double r, double g, double b){
+void graphic_draw_plus(unsigned int x,unsigned int y, double size, double r, double g, double b){
     (*ptcr)->set_line_width(size/3);
     (*ptcr)->set_source_rgb(r, g, b);
     (*ptcr)->move_to(x-size/2, y);
@@ -98,7 +98,7 @@ void graphic_draw_plus(double x,double y, double size, double r, double g, doubl
     (*ptcr)->line_to(x, y-size/2);
     (*ptcr)->stroke();
 }
-void graphic_draw_croix(double x, double y, double r, double g, double b ){
+void graphic_draw_croix(unsigned int x,unsigned int y, double r, double g, double b ){
     graphic_draw_carre_plein(x-3/2,y+3/2,1,r,g,b);
     graphic_draw_carre_plein(x-3/2,y-3/2,1,r,g,b);
     graphic_draw_carre_plein(x+3/2,y+3/2,1,r,g,b);
