+++ Code/graphic.h
@@ -5,9 +5,9 @@
 
 void graphic_set_context(const Cairo::RefPtr<Cairo::Context>& cr);
 void graphic_draw_window();
-void graphic_draw_carre_plein(double x,double y,double size, double r, double g, double b);
-void graphic_draw_carre_vide(double x, double y, double size, double r, double g, double b);
-void graphic_draw_losange(double x,double y,double r, double g, double b);
-void graphic_draw_plus(double x,double y, double size, double r, double g, double b);
-void graphic_draw_croix(double x, double y, double r, double g, double b );
+void graphic_draw_carre_plein(unsigned int x,unsigned int y,double size, double r, double g, double b);
+void graphic_draw_carre_vide(unsigned int x,unsigned int y, double size, double r, double g, double b);
+void graphic_draw_losange(unsigned int x,unsigned int y,double r, double g, double b);
+void graphic_draw_plus(unsigned int x,unsigned int y, double size, double r, double g, double b);
+void graphic_draw_croix(unsigned int x,unsigned int y, double r, double g, double b );
 #endif
